Fix Some Bugs

+++ app/Services/RoleService.php
@@ -3,6 +3,7 @@
 namespace App\Services;
 
 use DB;
+use Gate;
 use Spatie\Permission\Models\Role;
 
 /**
@@ -13,7 +14,7 @@ class RoleService
 
     public static function getAllRoles()
     {
-        $roles = Role::latest()->paginate(5);
+        $roles = Role::with("permissions")->paginate(5);
 
         return $roles;
     }
@@ -26,12 +27,13 @@ class RoleService
     public static function update($data, $id)
     {
         $role = Role::with("permissions")->findOrFail($id);
+        Gate::authorize("update", $role);
         $role->update($data);
         return $role;
     }
     public static function table($pageNumber)
     {
-        $roles = Role::latest()->paginate(5, ['*'], 'page', $pageNumber)->withPath(route("roles.index"))->onEachSide(0);
+        $roles = Role::with("permissions")->paginate(5, ['*'], 'page', $pageNumber)->withPath(route("roles.index"));
         return $roles;
     }
 
@@ -44,7 +46,8 @@ class RoleService
     public static function delete($id)
     {
 
-        $role = Role::findOrFail($id);
+        $role = Role::with("permissions")->findOrFail($id);
+        Gate::authorize("delete", $role);
         $role->delete();
         return $role;
     }
